Improve Detector.java scan method and Javadocs
Added Javadocs to the class and scan method.

Added toLowerCase call when scanning Jar files
for .class files to ensure capitalization does
not allow files to be missed.

Improved the scan method such that it returns
a boolean indicating whether an infection was
found or not. The Main.java class builds a list
of infected files, then outputs that at the end.

This improvement allows the same behavior but
enables the additional flexibility of Main.java
to run a scan directly from within Java. For
example, launchers or other applications wishing
to add automatic scanning.